Fourth part

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcUserRepository.java
@@ -4,7 +4,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.support.DataAccessUtils;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.ResultSetExtractor;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
@@ -15,30 +14,13 @@ import ru.javawebinar.topjava.model.User;
 import ru.javawebinar.topjava.repository.UserRepository;
 import ru.javawebinar.topjava.util.ValidationUtil;
 
-import java.util.*;
+import java.util.EnumSet;
+import java.util.List;
+import java.util.Set;
 
 @Repository
 @Transactional(readOnly = true)
 public class JdbcUserRepository implements UserRepository {
-    private final static ResultSetExtractor<List<User>> LIST_RESULT_SET_EXTRACTOR = rs -> {
-        List<User> users = new ArrayList<>();
-        while (rs.next()) {
-            User user = new User();
-            user.setId(rs.getInt("id"));
-            user.setName(rs.getString("name"));
-            user.setEmail(rs.getString("email"));
-            user.setPassword(rs.getString("password"));
-            user.setRegistered(rs.getDate("registered"));
-            user.setEnabled(rs.getBoolean("enabled"));
-            user.setCaloriesPerDay(rs.getInt("calories_per_day"));
-            EnumSet<Role> roles = EnumSet.noneOf(Role.class);
-            String role = rs.getString("roles");
-            if (role != null) roles.addAll(Arrays.stream(role.split(",")).map(Role::valueOf).toList());
-            user.setRoles(roles);
-            users.add(user);
-        }
-        return users;
-    };
     private final JdbcTemplate jdbcTemplate;
     private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
     private final SimpleJdbcInsert insertUser;
@@ -61,19 +43,22 @@ public class JdbcUserRepository implements UserRepository {
         if (user.isNew()) {
             Number newKey = insertUser.executeAndReturnKey(parameterSource);
             int userId = newKey.intValue();
-            this.insertBatchData(user, userId);
+            this.insertBatchRolesForUser(user, userId);
             user.setId(userId);
         } else {
             boolean isUpdatedUser = namedParameterJdbcTemplate.update("""
-                       UPDATE users SET name=:name, email=:email, password=:password, 
+                       UPDATE users SET name=:name, email=:email, password=:password,
                        registered=:registered, enabled=:enabled, calories_per_day=:caloriesPerDay WHERE id=:id
-                    """, parameterSource) == 0;
-            List<Role> rolesData = this.jdbcTemplate.queryForList("SELECT role FROM user_role WHERE user_id=?", Role.class, user.getId());
-            Set<Role> roles = !rolesData.isEmpty() ? EnumSet.copyOf(rolesData) : EnumSet.noneOf(Role.class);
-            boolean isUpdatedRole = !roles.equals(EnumSet.copyOf(user.getRoles()));
-            if (isUpdatedRole) {
-                jdbcTemplate.update("DELETE FROM user_role WHERE user_id=?", user.id());
-                this.insertBatchData(user, user.id());
+                    """, parameterSource) > 0;
+            boolean isUpdatedRole = false;
+            if (isUpdatedUser) {
+                List<Role> rolesData = this.jdbcTemplate.queryForList("SELECT role FROM user_role WHERE user_id=?", Role.class, user.getId());
+                Set<Role> roles = !rolesData.isEmpty() ? EnumSet.copyOf(rolesData) : EnumSet.noneOf(Role.class);
+                isUpdatedRole = !roles.equals(EnumSet.copyOf(user.getRoles()));
+                if (isUpdatedRole) {
+                    jdbcTemplate.update("DELETE FROM user_role WHERE user_id=?", user.id());
+                    this.insertBatchRolesForUser(user, user.id());
+                }
             }
             if (!isUpdatedUser && !isUpdatedRole)
                 return null;
@@ -81,12 +66,12 @@ public class JdbcUserRepository implements UserRepository {
         return user;
     }
 
-    private void insertBatchData(User user, int user1) {
+    private void insertBatchRolesForUser(User user, int userId) {
         jdbcTemplate.batchUpdate("INSERT INTO user_role (role, user_id) VALUES (?, ?)", user.getRoles(),
                 200,
                 (ps, argument) -> {
                     ps.setString(1, argument.name());
-                    ps.setInt(2, user1);
+                    ps.setInt(2, userId);
                 });
     }
